Added test case for Gaussian Core Model.
Added a test case that generates the phase diagram for the GC model.
This is the begining of a new features to be added in
the software, which means all test cases that are being added now, the building environment
and the way initial parameters are input into MD
will probably change.
In this commit the output paths for the data files were also changed to be those of the test.